use tighter types and add accessor for ADC value

+++ src/KTypeProbe.cpp
@@ -8,7 +8,7 @@ uint32_t overSampleRead(int N, uint32_t pin) {
     uint32_t value = 0;
     int i;
     for(i=0; i<N; i++ ) {
-    value += analogRead(pin);
+        value += analogRead(pin);
     }
     return value/i;
 }
@@ -19,7 +19,7 @@ milliCelcius_t KTypeProbe::readColdJunction()
 }
 
 /* Read the value of the attached K Type probe in ADC units */
-int32_t KTypeProbe::readProbe()
+uint32_t KTypeProbe::readProbe()
 {
     return overSampleRead(10, K_TYPE_PROBE);
 }
@@ -30,7 +30,7 @@ milliCelcius_t KTypeProbe::update()
     probeAdc = readProbe();
     probeOffset = CAL_TEMP_OFFSET_A * 1000 +
                   (CAL_TEMP_OFFSET_B - CAL_TEMP_OFFSET_A) * 1000 *
-                      (probeAdc - CAL_ADC_A) / (CAL_ADC_B - CAL_ADC_A);
+                      (((int32_t)probeAdc) - (int32_t)CAL_ADC_A) / ((int32_t)CAL_ADC_B - (int32_t)CAL_ADC_A);
     probe = probeOffset + coldJunction;
     return probe;
 };
